i80: Move scripts to root folder
