+++ firebase/functions/scripts/serve.js
@@ -1,5 +1,5 @@
 const { spawn } = require('child_process');
-const config = require('../../config/env.config');
+const config = require('../config/env.config');
 
 if (config.DEV_TOKEN) {
   spawn(
