+++ firebase/functions/scripts/shell.js
